% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifeexpgambia.R
\name{appGUI}
\alias{appGUI}
\title{This is an example function named 'data1'
data1, sampledata!
which prints 'data1, sampledata!'.
the following packages should be loaded to enable the shiny app view the three graphs through the fuction below.}
\usage{
appGUI()
}
\description{
This is an example function named 'data1'
data1, sampledata!
which prints 'data1, sampledata!'.
the following packages should be loaded to enable the shiny app view the three graphs through the fuction below.
}
